backend/requirements.txt
Pillow>=7.2.0
selenium>=3.141.0
webdrivermanager>=0.8.0